Add util for adding labels to k8s resource YAMLs
Summary:
we want to allow users to add custom labels on their pixie k8s resources.
we considered using the kustomize package to do these transformations, but after some investigation, the kustomize package isnt the easiest to use (unless you want to programatically write kustomize files, write the files to a directory, etc, someone links an example here https://github.com/kubernetes-sigs/kustomize/issues/142)
adding the labels to the YAMLs is easy enough. this diff contains the util for doing so.

next steps would be to allow the user to specify their custom labels in the `deploy` command on the CLI and write those flags to a config so that any update job that runs in the future can fetch them and apply them to future update yamls

Test Plan: deployed via CLI with some labels filled in, check that the cloudconn pod had the new labels applied

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7018

GitOrigin-RevId: eb40e772e2a027b6bc0b2b57810f75569bae5ee4