Add proto messages for PEM config updates
Summary:
this is part of our project for sending config updates to PEMs.
once the qb receives a PxL script with a mutation for a config update for a PEM, it should make a request to Metadata with the PEM's name and config change details.
The metadata then sends the config update to the PEM via NATS.
The proto definitions for that Metadata service RPC call is defined here, along with the message that the metadata sends to the PEM.

following the postgres alter system scheme: https://www.postgresql.org/docs/10/sql-altersystem.html we set keys/values one at a time.

also made the config settings arbitrary keys/values, although I could see some value in having more well-defined config settings. feedback on this would be appreciated

Test Plan: n/a

Reviewers: zasgar, philkuz, vihang, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7011

GitOrigin-RevId: ea41b17604c146b8295e37ae44e553a5c1657e36