Open the script editor when selecting the scratch pad script.
Summary: As the point of the scratch pad is to write something, it makes sense to open the editor and place the user in it.

Test Plan: Select the scratch pad script while the editor is closed. It should open. Known preexisting issue: the editor doesn't always correctly gain focus when it opens. See PC-473.

Reviewers: michelle, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-727

Differential Revision: https://phab.corp.pixielabs.ai/D7036

GitOrigin-RevId: c7f544d242427ff16d88f543efbe0f4c38ff2e13