Remove CLSX as a dependency in first-party code.
Summary:
The `react-virtualized` package still uses it internally, but it will no longer create an extra dependency (which can't be imported in tests) in our own code. This change replaces it with a short function that does the same thing, as well as a test for the same.
Additionally, this makes a minor edit to our `tsconfig.json` at both the top level and in `pixie-components`, to enable more recent JS features when compiling down to ES5.

Test Plan: Unit tests should pass; UI should have no user-visible changes whatsoever.

Reviewers: michelle, vihang, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7040

GitOrigin-RevId: 6a53873484b1f224d8e0d38d88003bdfd28a01d8