Instruct CodeCov not to analyze generated TypeScript definitions.
Summary: This directory is over 3300 lines of purely type definitions, and is tanking our coverage metric for the `src/ui/src` directory. Our coverage is indeed low, but it isn't 24% like CodeCov thinks right now. The `codecov.yml` file is merged at runtime into the upstream config (currently blank/defaults), so this does not clobber anything.

Test Plan: This is, unfortunately, a YOLO change: to test it, we have to land it, wait for Phab to propagate it to GitHub, and then see what happens when next CodeCov runs against `main`.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7043

GitOrigin-RevId: f5135c87e9313511c1de7686918b072f2f6f2557