Vz update job should deploy resources with user-specified annotations
Summary:
some users want to apply their custom annotations/labels to the resources we deploy on their cluster.
our resources are mostly deployed by the vz update job. this updates the job to take the user's custom labels (to be added in their cluster configmap by the CLI), and adds them to the deployed resources

Test Plan: made an rc, verified that resources deployed with the labels that i added to the configmap

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7045

GitOrigin-RevId: beb34ab7dae51832067810dcbfaf6ae8e25c3dc0