Update stan config to remove channel limits
Summary:
we currently have some limits set on the channels, we can set them to 0 to have "unlimited". i also reduced the max age on the channels so we store a little less data.
also updated the dev version of stan to have a smaller pv size

Test Plan: n/a

Reviewers: zasgar, vihang, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7052

GitOrigin-RevId: 99ef4f82141cd6f8baa4d45f2fd4daa32505ad38