Hookup Redis tracing
Summary:
Has been running this on gke:dev-cluster-yzhao (pl-pixies) for 12hours+:
* No pem crashes happens during the time
* CPU usage has been stable
* Query results have no visible errors.

Looks to me that this is OK to enable now.

Thanks to the recent change of inferring message type through the socket info, we did not need to
detect message content for the message type, which is critical to make it work for intercepting
an established connection.

In practice, inferring message type from socket info should be more reliable then through message
content, as Redis message content can be very generic. So I think this current mechanism is
appropriate as a long-term solution, until we discover series issues that necessitate examining
message content.

Test Plan: Jenkins & test cluster

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6975

GitOrigin-RevId: a891904b1b6cd34b9ac49b4710a06bc6b9f27233