pxconfig: mds should create mapping from podname->agentID
Summary:
when updating an agent config through pxl using pxconfig, the user needs to specify the pod name of the agent.
in order to forward the config update to the correct agent over NATS, mds needs to be able to convert the pod name to an agent ID.
in the current system, there is no clean way of doing that. we should just create a mapping of the agent's pod name -> its agent ID.

Test Plan: unit tests

Reviewers: zasgar, philkuz, nserrino, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7067

GitOrigin-RevId: faa7dd8b32e477ad23ca0a6c28022b294dd1ba64