Build our own curl image
Summary:
in our init containers, we were using some random docker image for `curl`. as @zasgar pointed out, this could be a security problem. and as @vihang pointed out, the image is simple enough that we can just build it ourselves.
this diff adds the logic for building that docker image, and also replaces the init container images with our new image.

Test Plan: ran skaffold with the new YAMLs

Reviewers: zasgar, vihang, philkuz, #engineering

Reviewed By: vihang, #engineering

Subscribers: zasgar, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7074

GitOrigin-RevId: 43ba6f3bc96e5b07a3b30df6c55b50f9356dc37b