UI coverage: Set up mocks and configuration to test Live View components.
Summary: By creating `breadcrumbs-test.tsx` to test just rendering the component, quite a bit of missing setup was discovered for live views. This diff builds that, including a single mock context (`live-context-mock.tsx`) to handle all of the common contexts used in these views. This adjusts a few exports (in particular converting one from default to named), thus editing so many files.

Test Plan: No-op change - just adds a test to verify that we can render a `<LiveViewBreadcrumbs />` component in Jest with the new setup.

Reviewers: michelle, zasgar, #engineering, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7078

GitOrigin-RevId: dac9cbaf8568016ce113a458c18473cdd109ce36