Make Pixie logo in sidebar a link back to the live view.
Summary: The rest of the menu item still opens and closes the sidebar. Also fixes a forward-ref-related error being logged to the console from the other Pixie icon in the breadcrumbs.

Test Plan: Open the sidebar from Admin, click on the logo. You should be returned to the live view.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-708

Differential Revision: https://phab.corp.pixielabs.ai/D7083

GitOrigin-RevId: 437f52bc26785deda31f7f282ef5af27792f7574