Better logic for clearing breadcrumb autocompletes when opening.
Summary: This was originally fixed by https://phab.corp.pixielabs.ai/D6960, but that turned out to be unreliable. This improves the logic.

Test Plan: Try opening the breadcrumb dropdowns in a variety of scenarios. Right after selecting a cluster or script, opening and immediately closing/opening again, after a script finishes loading, etc. No matter what, the input should blank and the list of completions should reflect that when opening the autocomplete.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-706

Differential Revision: https://phab.corp.pixielabs.ai/D7085

GitOrigin-RevId: b5d2b190af504bae3c6436a250eaf6594081f53e