Fix deploy --labels with true/false value
Summary: NR tried to deploy with our new --labels feature... only to find out it didn't work when they have a true/false as the label value. this is because trues/falses need to be quoted in the YAML file.

Test Plan: deployed through CLI with --labels=test=true, with/without extract_yamls, and also updated the cluster with labels to be sure

Reviewers: zasgar, #engineering, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7102

GitOrigin-RevId: 2410811f0bae4d1a2e4794c9f7e3122324b36f4b