PP-2396: Add table description to agent relation manager
Summary: We want to be able to query tables and their descriptions in a UDTF. This diff adds tracking of table descriptions in the RelationInfo class used by the agents to manage their relations. Next up, update the protobuf so that these descriptions are sent to the metadata service.

Test Plan: updated existing

Reviewers: philkuz, oazizi, #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: htroisi

JIRA Issues: PP-2396

Differential Revision: https://phab.corp.pixielabs.ai/D7138

GitOrigin-RevId: 9c1adb2cf933e21ae2c240b4f7ccb6be541ec31f