Add python protos for cloud api
Summary:
cloud api has a bunch of protos dependencies. :This converts them into the python format to use in the pixie python API. Most are necessary because they are needed by the VizierClusterInfo service. Not sure how to get around this.

Side note: also added a new proto compile function for just proto libs so we don't have to make grpc file when we only need proto. GRPC proto file is basically just an import grpc so is super useless.

Test Plan: n/a

Reviewers: vihang, zasgar, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7141

GitOrigin-RevId: 920983918bffb78a91921913ba1e8b8b6728472c