Remove unused metadata service API (SchemaByAgent)
Summary:
This has been sitting unimplemented for months so we should just remove it for now. I think a better solution for this API will be to give GetSchemas an optional agent ID list. Then, GetSchemasResponse will return a map of table to agents that contain that table. (this way we don't duplicate a full schema for each agent)

This API is unused and unimplemented so the API breaking change errors should be ok to ignore.

Test Plan: n/a

Reviewers: michelle, vihang, #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: vjain

Differential Revision: https://phab.corp.pixielabs.ai/D7153

GitOrigin-RevId: faf6ee060e1c0184b5ef7ab776ce84005f0c3088