PP-2396 Add GetTables UDTF to Carnot to fetch tables and their descriptions
Summary: Depends on D7154. Add a GetTables UDTF to Carnot which calls GetSchemas() in the metadata service (similar to GetTableSchemas()).

Test Plan: We didn't have a test for GetTableSchemas, likely due to the fact that that it would need to rely on a mock of the MDS and would likely not add much value beyond the template checks that are done during compilation, so I followed the same pattern here.

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Subscribers: vjain

JIRA Issues: PP-2396

Differential Revision: https://phab.corp.pixielabs.ai/D7155

GitOrigin-RevId: 1d24e0b6bc3611c0d16ef826569d6a41efc15f32