Add etcd implementation
Summary:
Adds a etcd implementation for use as a KVStore.
(Note: This is largely the same as the existing etcd implementation, but done
in a clean room to make switching to this easier. Also lines up with the pebbledb
implementation to have a common interface).

Test Plan: Added unit tests.

Reviewers: michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Subscribers: vjain

JIRA Issues: PP-2379

Differential Revision: https://phab.corp.pixielabs.ai/D7158

GitOrigin-RevId: e887543a6e412e2476c652c1a952c7da472f7df7