Add a DataStore interface
Summary:
While this doesn't quite follow the pattern of having the interface implemented at the consumer, it attempts to split up the interface into smaller logical pieces.

To avoid test dupe, and to ensure that both implementations behave similarly, I decided to combine the `etcd` and `pebbledb` tests into one test.
To run a single test on both implementations, we need a interface for the two datastores. Hence this does the same.

Test Plan: Refactored existing tests.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2376

Differential Revision: https://phab.corp.pixielabs.ai/D7125

GitOrigin-RevId: 311944a8fb67c7753278c8976012758a2ac5e482