PP-2370: Separate out Heartbeat Nack handler + set up piping for reregistration flow
Summary:
We want heartbeat nacks to have an option to tell the agent to reregister with the agent manager instead of just dying. This is to improve availability and service in the case the agent manager goes away and has to restart without context.

The plan will be to have the heartbeat nack message handler (now separate from the heartbeat handler) call a reregister hook in the event that it receives a heartbeat nack. It will still die if reregister is set to false (status quo).

Next up will be to rerfactor the agent registration logic so that it can also be used for agent reregistration. Then we will hook in reregistration into the nack handler.

Test Plan: existing/added

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: vihang, #engineering

Subscribers: vjain

JIRA Issues: PP-2370

Differential Revision: https://phab.corp.pixielabs.ai/D7195

GitOrigin-RevId: a01a69ef1b1ec3b2abb53ef943e80d5b5c83081d