Move uuidpb to public api
Summary: we need to make this public because we use it in some of our public apis

Test Plan:
tested by deploying to staging.
ran old, current vizier release. verified running a query with a uuid col (agent_status) works on skaffold cli and old cli release
ran vizier with skaffold. verified running a query with with a uuid col works on skaffold cli and old cli release

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7198

GitOrigin-RevId: f0408c91f436cee101b774a21034760385930e91