PP-2372 part 1 - Add agent handler for k8s updates
Summary: Currently we handle k8s updates from the HeartbeatAck from the metadata service. We want to move to handle them outside of heartbeats, because we want to support agents requesting a backlog of updates that they missed when that situation occurs. This diff sets up a basic handler for the k8s updates which replicates the functionality we have today (no requesting past updates) outside of the heartbeat handler. Next up, I will add the logic for rerequesting updates, and then, when the metadata service side of this is done, hook it in and remove the metadata side of the heartbeat handler.

Test Plan: unit, also added a fake for agent metadata state manager

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: vjain

JIRA Issues: PP-2372

Differential Revision: https://phab.corp.pixielabs.ai/D7257

GitOrigin-RevId: e6fc4bbbf2f399106ab8f5123e512658692a868d