PP-2370: integrate registration handler into manager.
Summary: Depends on D7215. This diff removes the agent registration logic from the manager.cc code and replaces it with the equivalent registration handler class. This is setting the piping up to support reregistration of agents when the agent manager loses context for which agents exist.

Test Plan: existing

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: vihang, zasgar, #engineering

Subscribers: vjain

JIRA Issues: PP-2370

Differential Revision: https://phab.corp.pixielabs.ai/D7216

GitOrigin-RevId: b43a0929b4c2918a105511e6bbb88b8c8fa9bbb7