TypeScript API library: pull GQL query definitions into `pixie-api` and start a `PixieAPIClient` definition.
Summary: This diff doesn't do much, but it affects enough places that wrapping it up in the next changes would make it hard to follow. See gigantic TODO in `api.ts` for descriptions of the remaining steps. Similar queries (like the various queries that requested the user table) were merged as part of this, as there are currently no fields that are expensive to fetch. If that changes, we'll need to consider a strategy for the consumer to specify which fields it wants to hydrate.

Test Plan: No-op from end user perspective; tests all still pass.

Reviewers: michelle, vihang, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-738

Differential Revision: https://phab.corp.pixielabs.ai/D7240

GitOrigin-RevId: dea525b05ffe1ef19950857a10f3ed7fea841f73