PP-2411: Add comment to LibuvTimer to call DeferredDelete on the scheduler before the destructor is invoked.
Summary: TSIA. This confused us because we thoguht the destructor would automatically disable the timer. However, in the LibuvScheduler, another step is needed. We don't want to put it in the destructor because it is beyond the responsibility/scope of the timer's destructor.

Test Plan: n/a

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: vjain

JIRA Issues: PP-2411

Differential Revision: https://phab.corp.pixielabs.ai/D7241

GitOrigin-RevId: 24a62a8068fa1ce6547f457983d1ded8a2108acd