PP-2370: Hook in logic in PEM to reregister agent and pause heartbeats until reregistration is complete
Summary:
The other diffs set this up in the agent registration and heartbeat side. The flow looks like this:
1. Agent receives HeartbeatNack w/ reregister flag set to true
2. Agent pauses heartbeats until reregistration is complete
3. Agent reregisters with metadata service.
4. When reregistration is complete, heartbeats resume.

We have unit tests for these components but will need the other side of the logic on the metadata service side to test it end to end.

Test Plan: existing unit

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: vihang, #engineering

Subscribers: vjain

JIRA Issues: PP-2370

Differential Revision: https://phab.corp.pixielabs.ai/D7242

GitOrigin-RevId: 21d832dd84fb0a58d993bb4b521b4c15e2c1102a