PP-2368: Add an interface for agent state manager and cleanup class
Summary:
Agent metadata state manager is a big, complicated class with code used in a lot of places. I did the following to make it easier to work with, since I was running into issues with it in my changes for PP-2368.

1. Give it an interface, so that tests can implement a fake and we don't need to spin up the entire thing for every dependency.
2. Clean up the public/private methods -- a lot of stuff was public that was only used inside the class.
3. Move out static methods that weren't really conceptually part of the class into standalone methods (or in one case a constant).

For some upcoming tests I will use a fake rather than for the full metadata service (such as my k8s update handler, which only cares if AddK8sUpdate was called with an update and doesn't need the full thing).

Test Plan: existing

Reviewers: vihang, zasgar, michelle, #engineering

Reviewed By: vihang, #engineering

Subscribers: vjain

JIRA Issues: PP-2368

Differential Revision: https://phab.corp.pixielabs.ai/D7244

GitOrigin-RevId: e86c6bd78e6edbdbb4cd125fac5f423155085922