Fix etcd wait container for very old Pixie deployments
Summary:
we recently added an etcd wait container for metadata, which pings the etcd endpoint to see if its healthy.
it takes the URL for the etcd endpoint from the cluster configmap... but very very old deployments of vizier do not have this configmap.
vizier deployments without the cluster configmap use the non-operator version of etcd by default, so should use the non-operator etcd URL.

Test Plan: ran skaffold with/without the configmap present

Reviewers: zasgar, vihang, philkuz, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7258

GitOrigin-RevId: 7bb4d7542a5735dfec1b64892f841ca4a110deaa