K8s updates: setDeletionTimestamp when k8s resource is deleted
Summary:
this fix got lost in the refactored version, so this diff adds it back in: https://phab.corp.pixielabs.ai/rPLM14577599e8be0938d8f2842a3dd1cc0a4510e21b
the deletionTimestamp is not necessarily set in the K8s API, however the agent downstream expects it to be set when the resource is actually deleted.
to address this, we have to look at the eventType from the k8s watcher. if it is a deletion event and the timestamp isnt set, we should just set it.

Test Plan: unit test

Reviewers: vihang, nserrino, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7273

GitOrigin-RevId: 6947880d51be745d2b3427b9ca7cd04d37ce8aa3