Add class CmdArgs to describe command arguments
Summary:
There is a change in formatting to remove the forced quotation marks.
That's a side effect of using Rapidjson, which itself adds its own quotation marks in formatting
array and map.

Fortunately, in the case of array response, the JSON format handles the detail.
In the case of other singular value response, the lost of quotation mark should be welcomed as it
becomes easier to query.

Test Plan: Jenkins

Reviewers: #engineering, jps, oazizi

Reviewed By: #engineering, oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7275

GitOrigin-RevId: 6885a695d3fb320ef7ea87b984a8e3270628242d