Add GetUpdatesToSend() to k8s UpdateProcessors
Summary:
once a k8s resource update is processed and stored in our data store, we need to send the update out to the relevant agents.
this diff adds the GetUpdatesToSend() function to the UpdateProcessors, which gets the array of updates that needs to be sent out, along with the IPs that the updates should be sent to.
- endpoints: we should send the update to all nodes that have a pod serving the endpoint. this is impossible to do with only the info provided in the endpoint update. as we receive pod updates, we need to create a map from podName->IP. now, when we receive the endpoint update, we can go from podName -> IP. the update we send out should also only contain references to pods running on that agent's node.
- service: these are covered by endpoints updates.
- pod: we need to send out the container updates and pod update to the relevant agent and the kelvin.
- node: we currently dont send out/the agent doesnt yet handle updates for nodes.
- namespace: we need to send the namespace update to all agents, including the kelvin. in order to get all IPs in our cluster, we create a map as we receive node updates.

the next diff will update the K8sMetadataHandler to call GetUpdatesToSend(), take those updates and publish them to NATS

Test Plan: unit tests

Reviewers: vihang, nserrino, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7276

GitOrigin-RevId: 6f7570c8e68fa4cc882994e8644dbf7e89d76b44