Adding results_test in place of some e2e tests
Summary:
e2e tests were pretty complicated. This simplifies a bit of the need by just testing results, which is where much of the complexity in the API is. Lots of this work is derived from
https://phab.corp.pixielabs.ai/D7274 where I had all the other machinery for the server in place. However, it was unnecessary for most of what we needed, so this approach basically cleans that up into a simple set of tests.

Note main doesn't support errors sent over the ExecuteScript stream. D7233 promises to do that, I'm probably gonna patch it in and try to land it.

Test Plan: It's a test and they pass.

Reviewers: vihang, zasgar, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7280

GitOrigin-RevId: 7276403c43fb8b53c3a2709dce2ea3f338822562