PP-2435 Map K8s resource versions to orderable resource versions
Summary:
It turns out that k8s resource versions are not comparable strings as we thought.
This means that we need to track our own resource versions, where these resource versions are easily comparable. to do so, the resource versions are numerical padded strings: for example 00000000000000000021
before, the GetStoredProtos/GetUpdatesToSend functions relied on the k8s resource version in the update. this change means that the k8s metadata handler is responsible for delegating the resource numbers, so those interfaces needed to be modified slightly.

Test Plan: unit test

Reviewers: nserrino, vihang, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7302

GitOrigin-RevId: b1d9f5f17356537d6d5dfc895e71462d4b7335d3