Get agent_topic_listener to ask unknown agents to re-register
Summary:
If we get a heartbeat from an agent we don't know about (this could
happen if the datastore we are using is only inmemory), just ask the agent
to reregister when sending the `nack`. This prevents agent death and loss of state.

When deleting known agents (this happens on heartbeat timeout), fire out
a `nack` without the re-register flag so that agent will die.

Test Plan: Updated unit test

Reviewers: nserrino, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7290

GitOrigin-RevId: 4379fdec47d1821d698b41770d05901e0cf09fc7