Update k8s metadata handler to store topic-specific vs unscoped resource updates
Summary:
as per comments in D7325, we want to avoid having to read the entire resource update history when fetching updates for a specific IP.
this diff updates the k8s metadata handler so that it writes a topic-specific resource update for updates that should go to specific nodes (pods, endpoints). other nodes will be written as unscoped updates.
now, when all resource updates are fetched for a specific IP, we fetch the topic-specific resource updates + the unscoped updates.

since we are now associating resource updates with specific topics in the datastore, we can store the partial resource update that is sent to the agents. before, we were storing the full resource update, which contains more data than we actually send to the agent. this was so we could reconstruct any agent-specific resource updates when needed.

Test Plan: unit tests

Reviewers: nserrino, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7328

GitOrigin-RevId: 0f6a0abf5409bcf9df49cd95a33343e60def0feb