Upgrade grpc to latest release
Summary:
TSIA. We no longer need to use the patched version.
This also required upgrading absl and sentencepiece.

Note: This was done in an attempt to fix the grpc build on macOS.
There are still some issues and grpc doesn't build on macOS yet.

Test Plan: Build and test all.

Reviewers: #engineering, zasgar, jamesbartlett

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D7343

GitOrigin-RevId: 0864e2bebc7000f11b399fd6900d965986a64e9f