PP-2368: Clean up agent manager (C++)
Summary:
There was some cruft built up in the agent manager that I was noticing as I was trying to add the k8s update handler to it. This diff addresses those issues:
1. redundant access to protected member variables via protected getter (moved member variables to private in accordance with Google C++ guidelines)
2. Unused constructors -> deleted
3. Many protected member variables are only used inside the class itself -> moved to private
4. Confusing logic with static member variables -> Moved to simple functions to reduce clutter in the already large agent manager class.
5. Moved StubGenerator to be its own function on agent::Manager to reduce clutter in constructor

These changes will make it slightly easier for me to add a new nats connector.

Test Plan: existing behavior should be the same, this is a pure refactor

Reviewers: vihang, zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Subscribers: vjain

JIRA Issues: PP-2368

Differential Revision: https://phab.corp.pixielabs.ai/D7348

GitOrigin-RevId: b13465e2dbe1443013219e00021bebf581468bfa