PP-2372: Add k8s update handler to agent manager.
Summary: Depends on D7348. This diff adds the k8s update handler to the agent manager's list of Vizier message handlers. This took some additional piping in comparison to other handlers because this handler listens to and writes to a different topic than the other handlers currently do.

Test Plan: we will need to run this with metadata service changes to really test this.

Reviewers: michelle, vihang, zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: vjain

JIRA Issues: PP-2372

Differential Revision: https://phab.corp.pixielabs.ai/D7350

GitOrigin-RevId: b445b23c4081f56678d7fe9a239bfb538185a92b