Add FetchFullResourceUpdate to metadata store
Summary:
the current FetchResourceUpdate function is for fetching partial topic-specific resource updates. this is used when agents/cloud requests missing metadata updates.
When MDS starts up, we will need to perform a "sync", which compares the current state of the k8s cluster, and the stored state in mds. this will use the FetchFullResourceUpdates function.

Test Plan: unit test

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7349

GitOrigin-RevId: 4e3939e1afe03df906e47c05de5df3a9fd306293