Add in missing call to reregister agent when heartbeat nack is received
Summary:
Agents would receive a heartbeat nack, disable heartbeats, but I forgot to add the reregister invocation in the agent handler, so they would just stay stuck, not heartbeating forever. This diff addresses that oversight.
This is related and possibly causal to the bug occuring on dev-cluster-stirling-perf.

Test Plan: n/a

Reviewers: vihang, michelle, #engineering

Reviewed By: vihang, michelle, #engineering

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7361

GitOrigin-RevId: 3b635bbac69a223a5ef8ee5fc7b5f3abeb44ac2d