PP-2372: Fix edge case with K8s update handler
Summary:
In the event that the k8s update handler made a request for missing metadata that looked like this: [4, 5)
and received a response that looked like this: updates:[], earliest: 6, latest: whatever,
The k8s update handler wasn't properly writing the backlog of updates that have version earlier than 6. This diff fixes that and adds a test.

Test Plan: added unit

Reviewers: vihang, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2372

Differential Revision: https://phab.corp.pixielabs.ai/D7365

GitOrigin-RevId: 56a681e7a56eaa749cc40f696203ba2d8c9cd3e4