Pod containers should also be terminated when the pod is terminated
Summary:
just like the other resource DeletionTimestamps, this StopTimestamp field in ContainerStatuses are optional fields from the k8s API. however, we depend on this field to be set to determine whether or not a resource has terminated. so, if we receive a k8s event of type Deleted, we should also mark the containers in the pod as deleted.
this is a fix that exists in the old metadata handler.

Test Plan: unit test

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7367

GitOrigin-RevId: 83a373e9e9ffb7432fdd531a6d1c45800f6c234e