Send metadatapb objs instead of k8s v1 objs on k8s metadata channel
Summary:
The k8s metadata handler used to accept k8s v1/runtime.Objects on the incoming update channel. This was largely because this was existing behavior of the old k8s metadata handler.
now that I'm refactoring the k8s metadata controller, which is the part of the pipeline responsible for actually watching the K8S API and putting the updates on the update channel, it seems to make more sense to have the k8s metadata handler take our metadatapb objects.
A large part of this is because of the syncing sequence (actual implmentation to come in a followup diff).
On startup, the controller will:
- make a request to the k8s API to get the current running state of the cluster
- make a request to the metadata store to get all of the stored updates, compiling the state of what was running at the time MDS stopped
- compare the two states, and for any resources that have since terminated, add a termination event on the update channel.
to do this last step with the old v1/runtime object type, we would have to convert the storepb.K8SResource message back to a runtime.Object type, shortly after which it would be converted back to a storepb.K8SResource message by the k8s handler. instead, we can just make the metadata controller responsible for converting it to our proto message before putting in on the channel.
this has the additional benefit that the k8s v1/runtime library is just imported in a single place.

Test Plan: existing unit tests

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7366

GitOrigin-RevId: 3102dab82876a0c408b5423034ce344c5400a418