Refactor metadata controller
Summary:
The metadata controller is responsible for listening to the K8s API and performing the sync sequence when MDS starts up.
I figured it would be cleaner to start in a separate file, rather than continue modifying the old metadata controller. the `k8s_metadata.go` should be replaced by `new_k8s_metadata.go` when the refactor is finished.
The largest change is how the sync is done. This syncing is done so that we don't have any zombie resources.
The sync sequence is as follows:
- Get the list of current running resources from the K8s API.
- Send those updates to the metadata handler. This is because we need to build up our NodeIP, PodName->PodIP maps, which we use to send our updates to the right IPs.
- Go through all stored resource updates. Find the resources that were still running at the time MDS stopped. Check if the resource is still running in K8s. If not, send the update event with type `Deleted`.
There's no way to get actual time at which the resource was deleted, so this is the best we can do. It is also safe to apply the current running updates first, since they should have no references to any deleted objects.

This diff uses `genny`, which is the closest we can get to generics in Go at the time. Hopefully when generics are actually implemented in Go, we can clean this all up.

Test Plan: unit test. also ran the new metadata controller in skaffold and confirmed events are coming through as expected

Reviewers: vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7369

GitOrigin-RevId: 69e0758283a54d1e319cc1c09224943dc0b202d9