Fix breadcrumb troll scroll behavior.
Summary:
Prevents the scroll view from jumping to the top all the time, or from jumping to the bottom if the mouse is too close to the bottom of the box, or being extremly slow to respond, or jumping up and down repeatedly if scrolling and moving the mouse at the same time.
The causes were twofold:
1) Excessive rerendering caused by non-memoized building of the breadcrumb contents.
2) Incomplete, rapidly-firing logic for bringing the currently selected element into view.

Test Plan: Run the local UI, and mess about with the breadcrumbs at the top of the live view. Scrolling and moving the mouse around within them should now behave substantially better.

Reviewers: #engineering, michelle, vihang

Reviewed By: #engineering, michelle

JIRA Issues: PC-767

Differential Revision: https://phab.corp.pixielabs.ai/D7368

GitOrigin-RevId: af2f585356b60ce823e05845d0985818a6fa5b31